requirements.txt
fastapi
uvicorn
yfinance # For share price
sqlalchemy # For DB interaction
loguru # For logging
python-jose[cryptography] 
bcrypt==4.0.1
passlib[bcrypt]
python-dotenv
gunicorn